refactor(voice/capture): extract LoopMixin (audio thread + consume loop) [T1.4 step 9b]
Extract the five tightly-coupled audio-thread + consume-loop
methods from ``voice/_capture_task.py`` into a new
``LoopMixin`` class at ``voice/capture/_loop_mixin.py`` per
CLAUDE.md anti-pattern #16:

* ``_audio_callback`` (~70 LOC) — PortAudio thread callback,
  T1.30-wrapped in ``try/except BaseException``.
* ``_enqueue`` (~6 LOC) — non-blocking queue insert with
  oldest-frame eviction.
* ``_check_sustained_underrun_rate`` (~80 LOC) — band-aid #9
  rolling-window xrun-fraction WARN.
* ``_consume_loop`` (~117 LOC) — main consumer task with
  band-aid #10 exponential-backoff reconnect on
  ``sd.PortAudioError``.
* ``_maybe_emit_heartbeat`` (~25 LOC) — periodic RMS / frame-
  count log.

Total: ~305 LOC migrated. ``_capture_task.py`` reduces from 1065
to 760 LOC — well under the CLAUDE.md anti-pattern #16 ~500 LOC
ceiling for the orchestration root (the remaining LOC is class
``__init__``, public lifecycle methods ``start`` / ``stop``,
status snapshots, and validation helpers — single-responsibility:
"compose the mixins + bind the public surface").

Class composition:
  AudioCaptureTask(EpochMixin, RingMixin, LifecycleMixin, RestartMixin)
    -> AudioCaptureTask(EpochMixin, RingMixin, LifecycleMixin, LoopMixin, RestartMixin)

LoopMixin sits BEFORE RestartMixin so RestartMixin's
``_audio_callback`` method-via-MRO stub is shadowed by the real
implementation in LoopMixin.

MRO TRAP — ``_reopen_stream_after_device_error`` lives on
RestartMixin (AFTER LoopMixin in MRO), but ``_consume_loop``
calls it via ``self._reopen_stream_after_device_error()``. A
naive ``def`` stub on LoopMixin would WIN MRO resolution and
shadow RestartMixin's real method (the stub returns ``None``,
making the reconnect path silently succeed without calling the
unified opener). Three ``TestExclusiveRestart`` tests caught
this immediately — the fix declares the cross-mixin reference
inside ``if TYPE_CHECKING:`` so the body is type-check-only and
not present at runtime, letting MRO fall through to
RestartMixin. Compatible-base-classes mypy strict check is
satisfied because the TYPE_CHECKING-block ``async def`` matches
RestartMixin's signature exactly. The pattern is documented in
the LoopMixin docstring for future cross-mixin refs.

Test-patch path migration (CLAUDE.md anti-pattern #20): 13
``patch("sovyx.voice._capture_task.time.monotonic", …)`` sites
in ``tests/unit/voice/test_capture_task.py`` migrated to
``patch("sovyx.voice.capture._loop_mixin.time.monotonic", …)``
in this same commit. Confirmed via grep + targeted re-run of
``TestConsumerLoopHeartbeatDriftT131`` and
``TestSustainedUnderrunWarnB9`` — all 13 sites resolve to the
new ``_loop_mixin.time`` reference and the patches still latch
the same code paths.

Mixin host-state declarations (24 attributes) cover everything
the five methods read/write so mypy strict accepts every
``self.X`` access without complaint. Two
``ChaosInjector`` / ``BackoffSchedule`` imports moved into
``if TYPE_CHECKING:`` blocks where they were no longer
runtime-needed after the consume-loop migration.

Gates: ruff + ruff format + mypy strict (429 files) + bandit
clean. Full pytest 11894 passed (zero delta vs step 9a).

Mission: docs-internal/missions/MISSION-voice-final-skype-grade-2026.md §Phase 1.T1.4 step 9b
