+++ src/sovyx/voice/capture/__init__.py
@@ -53,6 +53,7 @@ from sovyx.voice.capture._helpers import (
     _rms_db_int16,
 )
 from sovyx.voice.capture._lifecycle_mixin import LifecycleMixin
+from sovyx.voice.capture._loop_mixin import LoopMixin
 from sovyx.voice.capture._restart import (
     _LINUX_ALSA_HOST_API,
     _LINUX_SESSION_MANAGER_HOST_APIS,
@@ -85,6 +86,7 @@ __all__ = [
     "ExclusiveRestartResult",
     "ExclusiveRestartVerdict",
     "LifecycleMixin",
+    "LoopMixin",
     "RestartMixin",
     "RingMixin",
     "SessionManagerRestartResult",
